start and stop into own cog
